+++ noodles-bed/src/record.rs
@@ -11,7 +11,7 @@ pub use self::{builder::Builder, color::Color, name::Name, score::Score, strand:
 use std::{
     error,
     fmt::{self, Write},
-    num,
+    num::{self, NonZeroUsize},
     ops::Deref,
     str::FromStr,
 };
@@ -21,6 +21,8 @@ use noodles_core::Position;
 const DELIMITER: char = '\t';
 const MISSING_STRING: &str = ".";
 const MISSING_NUMBER: &str = "0";
+
+type Block = (usize, usize);
 
 #[derive(Clone, Debug, Eq, PartialEq)]
 struct StandardFields {
@@ -33,6 +35,7 @@ struct StandardFields {
     thick_start: Position,
     thick_end: Position,
     color: Option<Color>,
+    blocks: Vec<Block>,
 }
 
 impl StandardFields {
@@ -50,6 +53,7 @@ impl StandardFields {
             thick_start: start_position,
             thick_end: end_position,
             color: None,
+            blocks: Vec::new(),
         }
     }
 }
@@ -120,6 +124,7 @@ impl BedN<3> for Record<6> {}
 impl BedN<3> for Record<7> {}
 impl BedN<3> for Record<8> {}
 impl BedN<3> for Record<9> {}
+impl BedN<3> for Record<12> {}
 
 impl BedN<4> for Record<4> {}
 impl BedN<4> for Record<5> {}
@@ -127,26 +132,34 @@ impl BedN<4> for Record<6> {}
 impl BedN<4> for Record<7> {}
 impl BedN<4> for Record<8> {}
 impl BedN<4> for Record<9> {}
+impl BedN<4> for Record<12> {}
 
 impl BedN<5> for Record<5> {}
 impl BedN<5> for Record<6> {}
 impl BedN<5> for Record<7> {}
 impl BedN<5> for Record<8> {}
 impl BedN<5> for Record<9> {}
+impl BedN<5> for Record<12> {}
 
 impl BedN<6> for Record<6> {}
 impl BedN<6> for Record<7> {}
 impl BedN<6> for Record<8> {}
 impl BedN<6> for Record<9> {}
+impl BedN<6> for Record<12> {}
 
 impl BedN<7> for Record<7> {}
 impl BedN<7> for Record<8> {}
 impl BedN<7> for Record<9> {}
+impl BedN<7> for Record<12> {}
 
 impl BedN<8> for Record<8> {}
 impl BedN<8> for Record<9> {}
+impl BedN<8> for Record<12> {}
 
 impl BedN<9> for Record<9> {}
+impl BedN<9> for Record<12> {}
+
+impl BedN<12> for Record<12> {}
 
 impl<const N: u8> Record<N>
 where
@@ -423,6 +436,35 @@ where
     }
 }
 
+impl<const N: u8> Record<N>
+where
+    Self: BedN<12>,
+{
+    /// Returns the blocks (`[(blockStarts, blockSizes)]`).
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_bed as bed;
+    /// use noodles_core::Position;
+    ///
+    /// let blocks = vec![(0, 2)];
+    ///
+    /// let record = bed::Record::<12>::builder()
+    ///     .set_reference_sequence_name("sq0")
+    ///     .set_start_position(Position::try_from(8)?)
+    ///     .set_end_position(Position::try_from(13)?)
+    ///     .set_blocks(blocks.clone())
+    ///     .build()?;
+    ///
+    /// assert_eq!(record.blocks(), &blocks);
+    /// # Ok::<_, Box<dyn std::error::Error>>(())
+    /// ```
+    pub fn blocks(&self) -> &[Block] {
+        &self.standard_fields.blocks
+    }
+}
+
 impl fmt::Display for Record<3> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         format_bed_3_fields(f, self)?;
@@ -474,6 +516,14 @@ impl fmt::Display for Record<8> {
 impl fmt::Display for Record<9> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         format_bed_9_fields(f, self)?;
+        format_optional_fields(f, self.optional_fields())?;
+        Ok(())
+    }
+}
+
+impl fmt::Display for Record<12> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        format_bed_12_fields(f, self)?;
         format_optional_fields(f, self.optional_fields())?;
         Ok(())
     }
@@ -557,7 +607,10 @@ where
     write!(f, "{}", record.thick_end())
 }
 
-fn format_bed_9_fields(f: &mut fmt::Formatter<'_>, record: &Record<9>) -> fmt::Result {
+fn format_bed_9_fields<const N: u8>(f: &mut fmt::Formatter<'_>, record: &Record<N>) -> fmt::Result
+where
+    Record<N>: BedN<3> + BedN<4> + BedN<5> + BedN<6> + BedN<7> + BedN<8> + BedN<9>,
+{
     format_bed_8_fields(f, record)?;
 
     f.write_char(DELIMITER)?;
@@ -567,6 +620,36 @@ fn format_bed_9_fields(f: &mut fmt::Formatter<'_>, record: &Record<9>) -> fmt::R
     } else {
         f.write_str(MISSING_NUMBER)
     }
+}
+
+fn format_bed_12_fields(f: &mut fmt::Formatter<'_>, record: &Record<12>) -> fmt::Result {
+    format_bed_9_fields(f, record)?;
+
+    let blocks = record.blocks();
+
+    write!(f, "{}{}", DELIMITER, blocks.len())?;
+
+    f.write_char(DELIMITER)?;
+
+    for (i, (_, size)) in blocks.iter().enumerate() {
+        if i > 0 {
+            f.write_char(',')?;
+        }
+
+        write!(f, "{}", size)?;
+    }
+
+    f.write_char(DELIMITER)?;
+
+    for (i, (start, _)) in blocks.iter().enumerate() {
+        if i > 0 {
+            f.write_char(',')?;
+        }
+
+        write!(f, "{}", start)?;
+    }
+
+    Ok(())
 }
 
 fn format_optional_fields(
@@ -618,6 +701,18 @@ pub enum ParseError {
     MissingColor,
     /// The color is invalid.
     InvalidColor(color::ParseError),
+    /// The block count is missing.
+    MissingBlockCount,
+    /// The block count is invalid.
+    InvalidBlockCount(num::ParseIntError),
+    /// The block sizes are missing.
+    MissingBlockSizes,
+    /// A block size is invalid.
+    InvalidBlockSize(num::ParseIntError),
+    /// The block starts are missing.
+    MissingBlockStarts,
+    /// A block start is invalid.
+    InvalidBlockStart(num::ParseIntError),
 }
 
 impl error::Error for ParseError {}
@@ -642,6 +737,12 @@ impl fmt::Display for ParseError {
             Self::InvalidThickEnd(e) => write!(f, "invalid thick end: {}", e),
             Self::MissingColor => f.write_str("missing color"),
             Self::InvalidColor(e) => write!(f, "invalid color: {}", e),
+            Self::MissingBlockCount => f.write_str("missing block count"),
+            Self::InvalidBlockCount(e) => write!(f, "invalid block count: {}", e),
+            Self::MissingBlockSizes => f.write_str("missing block sizes"),
+            Self::InvalidBlockSize(e) => write!(f, "invalid block size: {}", e),
+            Self::MissingBlockStarts => f.write_str("missing block starts"),
+            Self::InvalidBlockStart(e) => write!(f, "invalid block start: {}", e),
         }
     }
 }
@@ -723,6 +824,17 @@ impl FromStr for Record<9> {
     }
 }
 
+impl FromStr for Record<12> {
+    type Err = ParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let mut fields = s.split(DELIMITER);
+        let standard_fields = parse_bed_12_fields(&mut fields)?;
+        let optional_fields = parse_optional_fields(&mut fields);
+        Ok(Self::new(standard_fields, optional_fields))
+    }
+}
+
 fn parse_bed_3_fields<'a, I>(fields: &mut I) -> Result<StandardFields, ParseError>
 where
     I: Iterator<Item = &'a str>,
@@ -781,6 +893,15 @@ where
 {
     let mut standard_fields = parse_bed_8_fields(fields)?;
     standard_fields.color = parse_color(fields)?;
+    Ok(standard_fields)
+}
+
+fn parse_bed_12_fields<'a, I>(fields: &mut I) -> Result<StandardFields, ParseError>
+where
+    I: Iterator<Item = &'a str>,
+{
+    let mut standard_fields = parse_bed_9_fields(fields)?;
+    standard_fields.blocks = parse_blocks(fields)?;
     Ok(standard_fields)
 }
 
@@ -899,6 +1020,42 @@ where
             MISSING_NUMBER => Ok(None),
             _ => s.parse().map(Some).map_err(ParseError::InvalidColor),
         })
+}
+
+fn parse_blocks<'a, I>(fields: &mut I) -> Result<Vec<Block>, ParseError>
+where
+    I: Iterator<Item = &'a str>,
+{
+    const LIST_DELIMITER: char = ',';
+
+    let len = fields
+        .next()
+        .ok_or(ParseError::MissingBlockCount)
+        .and_then(|s| {
+            s.parse::<NonZeroUsize>()
+                .map_err(ParseError::InvalidBlockCount)
+        })
+        .map(usize::from)?;
+
+    let raw_sizes = fields
+        .next()
+        .ok_or(ParseError::MissingBlockSizes)
+        .map(|s| s.splitn(len, LIST_DELIMITER))?;
+
+    let raw_starts = fields
+        .next()
+        .ok_or(ParseError::MissingBlockStarts)
+        .map(|s| s.splitn(len, LIST_DELIMITER))?;
+
+    let mut blocks = Vec::with_capacity(len);
+
+    for (raw_start, raw_size) in raw_starts.zip(raw_sizes) {
+        let start = raw_start.parse().map_err(ParseError::InvalidBlockStart)?;
+        let size = raw_size.parse().map_err(ParseError::InvalidBlockSize)?;
+        blocks.push((start, size));
+    }
+
+    Ok(blocks)
 }
 
 fn parse_optional_fields<'a, I>(fields: &mut I) -> OptionalFields
@@ -1079,6 +1236,38 @@ mod tests {
     }
 
     #[test]
+    fn test_fmt_for_record_12() -> Result<(), noodles_core::position::TryFromIntError> {
+        let start = Position::try_from(8)?;
+        let end = Position::try_from(13)?;
+
+        let mut standard_fields = StandardFields::new("sq0", start, end);
+        standard_fields.blocks = vec![(0, 2)];
+        let record: Record<12> = Record::new(standard_fields, OptionalFields::default());
+        assert_eq!(record.to_string(), "sq0\t7\t13\t.\t0\t.\t7\t13\t0\t1\t2\t0");
+
+        let mut standard_fields = StandardFields::new("sq0", start, end);
+        standard_fields.blocks = vec![(0, 2), (3, 1)];
+        let record: Record<12> = Record::new(standard_fields, OptionalFields::default());
+        assert_eq!(
+            record.to_string(),
+            "sq0\t7\t13\t.\t0\t.\t7\t13\t0\t2\t2,1\t0,3"
+        );
+
+        let mut standard_fields = StandardFields::new("sq0", start, end);
+        standard_fields.blocks = vec![(0, 2)];
+        let record: Record<12> = Record::new(
+            standard_fields,
+            OptionalFields::from(vec![String::from("ndls")]),
+        );
+        assert_eq!(
+            record.to_string(),
+            "sq0\t7\t13\t.\t0\t.\t7\t13\t0\t1\t2\t0\tndls"
+        );
+
+        Ok(())
+    }
+
+    #[test]
     fn test_from_str_for_record_3() -> Result<(), noodles_core::position::TryFromIntError> {
         let actual = "sq0\t7\t13".parse::<Record<3>>();
 
@@ -1183,6 +1372,24 @@ mod tests {
         standard_fields.thick_start = start;
         standard_fields.thick_end = end;
         standard_fields.color = Some(Color::RED);
+
+        let expected = Ok(Record::new(standard_fields, OptionalFields::default()));
+
+        assert_eq!(actual, expected);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_from_str_for_record_12() -> Result<(), noodles_core::position::TryFromIntError> {
+        let actual = "sq0\t7\t13\t.\t0\t.\t7\t13\t0\t1\t2\t0".parse::<Record<12>>();
+
+        let start = Position::try_from(8)?;
+        let end = Position::try_from(13)?;
+        let mut standard_fields = StandardFields::new("sq0", start, end);
+        standard_fields.thick_start = start;
+        standard_fields.thick_end = end;
+        standard_fields.blocks = vec![(0, 2)];
 
         let expected = Ok(Record::new(standard_fields, OptionalFields::default()));